Implements remote Map/Sketch support over Akka actors.
Users of the client side actor (like the soon to be developed RemoteDataSet),
use the ask pattern to issue {Map,Sketch}Operations to the client actor and
get back an observable. Once a subscribe is called on the returned observable,
the {Map,Sketch} command is fired off to the remote actor and executed, and
the results of the onNext(), onCompleted(), and onError() invocations on
the remote side are streamed to the client actor (and subsequently, to the
client-side observer). Unsubscription and cancellation work is still in progress.